requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.1.1
python-telegram-bot==20.7
sniffio==1.3.1
typing_extensions==4.15.0
Werkzeug==3.1.3
